Add JWT authentication to the Delete_order endpoint

+++ Routes/orders.py
@@ -15,7 +15,7 @@ async def create_shopping_cart(auth_token: Annotated[str, Header()]):
         payload = jwt.decode(auth_token, "secret", algorithms="HS256")
         user_id: str = payload.get("user_id")
         if user_id is None:
-            raise HTTPException(status_code=400, detail="The provided token does not have user id, please try logging in again.")
+            raise HTTPException(status_code=401, detail="The provided token does not have user id, please try logging in again.")
     except JWTError:
         raise HTTPException(status_code=401, detail="Could not validate credentials")
 
@@ -30,11 +30,22 @@ async def create_shopping_cart(auth_token: Annotated[str, Header()]):
    
 # DELETE /v1/orders/uuid
 @router.delete("/v1/orders/{uuid}")
-def delete_shopping_cart(uuid: str):
+def delete_shopping_cart(uuid: str, auth_token: Annotated[str, Header()]):
+    try:
+        payload = jwt.decode(auth_token, "secret", algorithms="HS256")
+        user_id: str = payload.get("user_id")
+        if user_id is None:
+            raise HTTPException(status_code=401, detail="The provided token does not have user id, please try logging in again.")
+    except JWTError:
+        raise HTTPException(status_code=401, detail="Could not validate credentials")
+    
     existing_item = ddb.get_item(Key={"cart_id": uuid})
     
     if not existing_item.get("Item"):
         raise HTTPException(status_code=404, detail="Shopping cart was not found for this user")
+    
+    if existing_item.get("Item").get("owner_id") != user_id:
+        raise HTTPException(status_code=403, detail="You are not authorized to delete this shopping cart")
     
     try:
         ddb.delete_item(Key={"cart_id": uuid})
